move dropzone element
