+++ src/Umbraco.Web.UI.Client/src/packages/media/media/collection/media-collection.element.ts
@@ -1,7 +1,7 @@
 import { UMB_MEDIA_ENTITY_TYPE, UMB_MEDIA_ROOT_ENTITY_TYPE } from '../entity.js';
 import { UMB_MEDIA_WORKSPACE_CONTEXT } from '../workspace/media-workspace.context-token.js';
 import type { UmbDropzoneSubmittedEvent } from '../dropzone/dropzone-submitted.event.js';
-import type { UmbDropzoneElement } from '../dropzone/dropzone.element.js';
+import type { UmbDropzoneElement } from '../dropzone/index.js';
 import { UMB_MEDIA_COLLECTION_CONTEXT } from './media-collection.context-token.js';
 import { customElement, html, ref, state, when } from '@umbraco-cms/backoffice/external/lit';
 import { UmbCollectionDefaultElement } from '@umbraco-cms/backoffice/collection';
