+++ src/Umbraco.Web.UI.Client/src/packages/media/media/dropzone/components/index.ts
@@ -1,0 +1,1 @@
+export * from './dropzone.element.js';
