+++ src/Umbraco.Web.UI.Client/src/packages/media/media/dropzone/index.ts
@@ -1,4 +1,4 @@
-export * from './dropzone.element.js';
+export * from './components/index.js';
 export * from './dropzone-manager.class.js';
 export * from './dropzone-submitted.event.js';
 export * from './dropzone-change.event.js';
